Use gpt-5.1-codex-mini, adjust credit system

+++ cloud/usage/types.ts
@@ -3,7 +3,7 @@ import { z } from 'zod';
 export const usageSchema = z.object({
   billingPeriodStartDate: z.date(),
   usage: z.object({
-    aiCredits: z.number().int().nonnegative(),
+    aiCredits: z.number().nonnegative(),
     queryDurationMilliseconds: z.number().int().nonnegative(),
     queryResponseBytes: z.number().int().nonnegative(),
   }),
